New Patch
play_card function (but has to be modified)
Two new functions to "lista_dinamica.h", listad_peca_pos that go through a list and find a item on a X position. listad_soma that go through the list and return the sum of all pieces. (can be improved)
CheckWin added but it has to test
Game Logic in the lib